src: clean-cut move to release/10.3.0
Taken from: FreeBSD
Commit ref: b62280e683e2

+++ contrib/ntp/libntp/ntp_worker.c
@@ -267,7 +267,7 @@ process_blocking_resp(
 void
 harvest_blocking_responses(void)
 {
-	size_t		idx;
+	int		idx;
 	blocking_child*	cp;
 	u_int		scseen, scdone;
